@@ -3,16 +3,20 @@
   "name": "react-typescript-starter-pack",
   "version": "0.1.0",
   "dependencies": {
+    "@reduxjs/toolkit": "^1.9.1",
     "@testing-library/jest-dom": "^5.12.0",
-    "@testing-library/react": "^11.2.6",
+    "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
     "@types/jest": "^26.0.23",
     "@types/node": "^12.20.10",
-    "@types/react": "^17.0.4",
-    "@types/react-dom": "^17.0.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "4.0.3",
+    "axios": "^1.2.1",
+    "classnames": "^2.3.2",
+    "core-js": "^3.27.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.5",
+    "react-scripts": "^4.0.3",
+    "redux-persist": "^6.0.0",
     "typescript": "^4.2.4",
     "typescript-eslint-parser": "^22.0.0",
     "web-vitals": "^1.1.1"
@@ -46,6 +50,8 @@
   },
   "devDependencies": {
     "@mate-academy/eslint-config-react-typescript": "^0.0.4",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.10",
     "eslint": "^7.25.0",
     "gh-pages": "^3.1.0",
     "node-sass": "^5.0.0"
